backend-python/app.py
"""
Flask backend for RAG Knowledge Retrieval System
"""
import os
import logging
from flask import Flask, request, jsonify
from flask_cors import CORS
from dotenv import load_dotenv

load_dotenv()

app = Flask(__name__)
CORS(app)

logger = logging.getLogger(__name__)


@app.route('/health', methods=['GET'])
def health():
    """Health check endpoint"""
    return jsonify({
        'status': 'healthy',
        'service': 'RAG Knowledge Retrieval System'
    }), 200


@app.route('/api/query', methods=['POST'])
def query():
    """Query the knowledge base"""
    try:
        data = request.get_json()
        query_text = data.get('query', '')
        k = data.get('k', 5)
        
        if not query_text:
            return jsonify({'error': 'Query is required'}), 400
        
        # TODO: Integrate with knowledge retrieval system
        results = {
            'query': query_text,
            'results': [],
            'count': 0
        }
        
        return jsonify(results), 200
    except Exception as e:
        logger.error(f"Error processing query: {e}")
        return jsonify({'error': str(e)}), 500


@app.route('/api/ingest', methods=['POST'])
def ingest():
    """Ingest documents"""
    try:
        files = request.files.getlist('files')
        
        if not files:
            return jsonify({'error': 'No files provided'}), 400
        
        # TODO: Process files and store embeddings
        
        return jsonify({
            'message': f'Ingested {len(files)} files',
            'count': len(files)
        }), 200
    except Exception as e:
        logger.error(f"Error ingesting files: {e}")
        return jsonify({'error': str(e)}), 500


@app.route('/api/status', methods=['GET'])
def status():
    """Get system status"""
    return jsonify({
        'status': 'operational',
        'documents_indexed': 0,
        'embeddings_stored': 0
    }), 200


if __name__ == '__main__':
    port = int(os.getenv('FLASK_PORT', 5000))
    debug = os.getenv('FLASK_ENV') == 'development'
    app.run(host='0.0.0.0', port=port, debug=debug)
