requirements.txt
openai==1.3.0
pinecone-client==3.0.2
faiss-cpu==1.7.4
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.0
python-dotenv==1.0.0
pydantic==2.0.0
numpy==1.24.0
pandas==2.0.0
